Program Description:
You have a bag containing some black beans and some white beans.  You also have an unlimited supply of black beans.  You are going to remove beans from the bag according to the following scheme. 

Task
If
Remove 2 beans from the bag. 

They are both black return one of them to the bag and discard the other.

One is black and the other white return the white bean to the bag and discard the black bean

Both are white discard both and put a black bean in the bag

What color is the last bean?

You decide to write a program that you can run for 50 simulations.  Start with a bag of 10 beans and continue running the program incrementing the number of beans in the bag by one until you have tested for a bag with 60 beans.  Fill the bag with random colors of beans and with draw the beans from the bag on a random basis, as well.

Run the simulation and print out four columns of numbers:  1) Number of beans 2) Number of Black Beans 3) number of White Beans 4) color of last bean.

Analyze your results and draw a conclusion from the data so that you can predict which color bean will be left in the bag.  Can you state mathematically why this conclusion is true?  Do this in your program documentation.

Use the ArrayList Class to implement this program.  Remove beans from random positions within the array and add beans back in a different random position.  

Hint:  Remember to generate your random number based on the size of the array.

Statements Required: random numbers, decision making, loops, ArrayList, methods, class

Sample Output:

Beans	Black	White	Last
10	6	4	White
11	7	4	White
12	5	7	Black	
13	6	7	Black
14	7	7	Black
15	7	8	Black
.
.
.
60	24	36	Black